move rosie/core/tests to tests/core and improve tests
