--- src/rosie/core/tests/__init__.py
@@ -1,3 +1,0 @@
-import os
-
-FIXTURES_DIR = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'fixtures')
