+++ tests/chamber_of_deputies/test_adapter.py
@@ -14,9 +14,9 @@ from rosie.chamber_of_deputies.adapter import COLUMNS as ADAPTER_COLUMNS
 
 class TestAdapter(TestCase):
     def setUp(self):
-        from . import get_fixtures_dir
+        from ..utils import get_fixtures_dir
         self.temp_path = mkdtemp()
-        self.fixtures_path = get_fixtures_dir()
+        self.fixtures_path = get_fixtures_dir(__file__)
         copies = (
             ('companies.xz', subject_class.COMPANIES_DATASET),
             ('reimbursements.xz', 'reimbursements.xz')
