+++ tests/chamber_of_deputies/test_meal_price_outlier_classifier.py
@@ -11,8 +11,8 @@ from rosie.chamber_of_deputies.classifiers.meal_price_outlier_classifier import 
 
 class TestMealPriceOutlierClassifier(TestCase):
     def setUp(self):
-        from . import get_fixtures_dir
-        FIXTURES_DIR = get_fixtures_dir()
+        from ..utils import get_fixtures_dir
+        FIXTURES_DIR = get_fixtures_dir(__file__)
         the_file = os.path.join(FIXTURES_DIR, 'meal_price_outlier_classifier.csv')
         self.dataset = pd.read_csv(the_file,
                                    dtype={'recipient_id': np.str})
