+++ tests/chamber_of_deputies/test_traveled_speeds_classifier.py
@@ -11,8 +11,8 @@ from rosie.chamber_of_deputies.classifiers.traveled_speeds_classifier import Tra
 
 class TestTraveledSpeedsClassifier(TestCase):
     def setUp(self):
-        from . import get_fixtures_dir
-        FIXTURES_DIR = get_fixtures_dir()
+        from ..utils import get_fixtures_dir
+        FIXTURES_DIR = get_fixtures_dir(__file__)
         the_file = os.path.join(FIXTURES_DIR, 'traveled_speeds_classifier.csv')
         self.dataset = pd.read_csv(the_file,
                                    dtype={'recipient_id': np.str})
