+++ tests/cli/test_rosie.py
@@ -8,5 +8,12 @@ def test_main():
     runner = CliRunner()
     result = runner.invoke(main, [])
 
-    assert result.output == '()\n'
+    assert result.output == """Usage: root [OPTIONS] COMMAND [ARGS]...
+
+Options:
+  --help  Show this message and exit.
+
+Commands:
+  chamber_of_deputies  serenata data path
+"""
     assert result.exit_code == 0
