+++ tests/core/fixtures/invalid_cnpj_cpf_classifier.csv
