+++ tests/utils.py
@@ -1,0 +1,5 @@
+def get_fixtures_dir(file):
+    import os
+
+    FIXTURES_DIR = os.path.join(os.path.dirname(os.path.abspath(file)), 'fixtures')
+    return FIXTURES_DIR
